Move some utility methods out of OmniHID
